Create logon_process_trusted.yml
This event indicates that a logon process has registered with the Local Security Authority (LSA). At the technical level, the event does not come from the registration of a trusted logon process, but from a confirmation that the process is a trusted logon process. If it is a trusted logon process, the event generates.